move artist page to subfolder
